+++ packages/solid/src/validation/custom.ts
@@ -12,8 +12,10 @@ export function custom<TFieldValue extends FieldValue>(
   requirement: (value: TFieldValue | undefined) => boolean | Promise<boolean>,
   error: string
 ): (value: TFieldValue) => Promise<string> {
-  return async (value: TFieldValue) => {
-    const result = await requirement(value);
-    return !result ? error : '';
-  };
+  return async (value: TFieldValue) =>
+    (value instanceof FileList || Array.isArray(value)
+      ? value.length
+      : value || value === 0) && !(await requirement(value))
+      ? error
+      : '';
 }