+++ packages/solid/src/validation/maxNumber.ts
@@ -11,5 +11,5 @@ export function maxNumber(
   error: string
 ): (value: number | null | undefined) => string {
   return (value: number | null | undefined) =>
-    (value || 0) > requirement ? error : '';
+    (value || value === 0) && value > requirement ? error : '';
 }