+++ packages/solid/CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Change type of `value` validation function
+- Add support for validation of optional fields (issue #5)
 
 ## v0.8.4 (November 29, 2022)
 
-- Fix circular reference type error of issue #4
+- Fix circular reference type error (issue #4)
 
 ## v0.8.3 (November 28, 2022)